requirements.txt
numpy == 1.24.4
psycopg[binary] == 3.1.18
bcrypt == 4.1.2
fastapi == 0.104.1
uvicorn[standard] == 0.24.0
python-jose[cryptography] == 3.3.0
python-multipart == 0.0.6
python-dotenv == 1.0.0